+++ packages/icons/src/icons/play-icon.js
@@ -1,15 +1,21 @@
 import React from "react";
+import PropTypes from "prop-types";
+import { clean } from "@times-components/utils";
 
-const PlayIcon = ({fill}) => (
+const PlayIcon = ({
+  fill = "black",
+  width = 24,
+  height = 25,
+  title = "Play Icon"
+}) => (
   <svg
-    width="24"
-    height="25"
-    viewBox="0 0 24 25"
-    fill={fill}
-    xmlns="http://www.w3.org/2000/svg"
-    aria-label="Play Icon"
+    aria-label="play-icon"
     role="img"
+    viewBox="0 0 24 25"
+    {...clean({ width, height, title })}
+    xmlns="http://www.w3.org/2000/svg"
   >
+    <title>{title}</title>
     <mask
       id="mask0_4528_3091"
       style={{ maskType: "alpha" }}
@@ -19,10 +25,7 @@ const PlayIcon = ({fill}) => (
       width="24"
       height="25"
     >
-      <path
-        d="M8 5.66699V19.667L19 12.667L8 5.66699Z"
-        fill={fill}
-      />
+      <path d="M8 5.66699V19.667L19 12.667L8 5.66699Z" fill={fill} />
     </mask>
 
     <g mask="url(#mask0_4528_3091)">
@@ -30,5 +33,19 @@ const PlayIcon = ({fill}) => (
     </g>
   </svg>
 );
+
+PlayIcon.propTypes = {
+  fill: PropTypes.string,
+  width: PropTypes.number,
+  height: PropTypes.number,
+  title: PropTypes.string
+};
+
+PlayIcon.defaultProps = {
+  fill: "black",
+  width: 24,
+  height: 25,
+  title: "Play Icon"
+};
 
 export default PlayIcon;