+++ packages/save-and-share-bar/src/tracking/with-track-events.js
@@ -58,6 +58,26 @@ export default Component =>
           articleHeadline
         }),
         trackingName: "ArticleShareOnTwitter"
+      },
+      {
+        actionName: "Pressed",
+        eventName: "onShareBarExpend",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline,
+          event_social_action: "share start"
+        }),
+        trackingName: "ArticleShareOnExpend"
+      },
+      {
+        actionName: "Pressed",
+        eventName: "onShareBarClose",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline,
+          event_social_action: "share start"
+        }),
+        trackingName: "ArticleShareOnClose"
       }
     ]
   });