+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.92.5](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.92.4...@times-components/article-main-comment@2.92.5) (2024-11-19)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [2.92.4](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.92.3...@times-components/article-main-comment@2.92.4) (2024-11-14)
 
 **Note:** Version bump only for package @times-components/article-main-comment